Move context resolver to analyzer
