+++ src/main/java/de/chojo/repbot/analyzer/ContextResolver.java
@@ -2,9 +2,11 @@ package de.chojo.repbot.util;
 
 import de.chojo.jdautil.parsing.Verifier;
 import de.chojo.repbot.data.VoiceData;
+import de.chojo.repbot.data.wrapper.GuildSettings;
 import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.entities.Message;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 import javax.sql.DataSource;
 import java.time.Instant;
@@ -29,15 +31,14 @@ public class ContextResolver {
      * Only members which have written in the last 100 messages which are not older than the max history and are not
      * send before the first message of the message author are returned
      *
-     * @param message       message to determine channel, author and start time
-     * @param maxHistoryAge max age of analyzed messages. may be overriden by first message of author if this is more
-     *                      recent.
+     * @param message  message to determine channel, author and start time
+     * @param settings setting sof the guild.
      * @return list of members which have written in this channel
      */
     @NotNull
-    public Set<Member> getChannelContext(Message message, int maxHistoryAge) {
+    public Set<Member> getChannelContext(Message message, GuildSettings settings) {
         var history = message.getChannel().getHistoryBefore(message, 50).complete();
-        var maxAge = Instant.now().minus(maxHistoryAge, ChronoUnit.MINUTES);
+        var maxAge = Instant.now().minus(settings == null ? Long.MAX_VALUE : settings.maxMessageAge() , ChronoUnit.MINUTES);
         var retrievedHistory = new ArrayList<>(history.getRetrievedHistory());
         // add user message
         retrievedHistory.add(message);
@@ -65,20 +66,21 @@ public class ContextResolver {
                 .collect(Collectors.toCollection(HashSet::new));
     }
 
-    public Set<Member> getVoiceContext(Message message, int maxAge) {
+    public Set<Member> getVoiceContext(Message message, @Nullable  GuildSettings settings) {
         Set<Member> members = new LinkedHashSet<>();
         var voiceState = message.getMember().getVoiceState();
         if (voiceState.inVoiceChannel()) members.addAll(voiceState.getChannel().getMembers());
-        var pastUser = voiceData.getPastUser(message.getAuthor(), message.getGuild(), maxAge, 10);
+        var pastUser = voiceData.getPastUser(message.getAuthor(), message.getGuild(),
+                settings == null ? 0 : settings.maxMessageAge(), 10);
         return pastUser.stream()
                 .map(id -> message.getGuild().retrieveMemberById(id).onErrorMap(throwable -> null).complete())
                 .collect(Collectors.toCollection(() -> members));
     }
 
-    public Set<Member> getCombinedContext(Message message, int maxAge) {
+    public Set<Member> getCombinedContext(Message message, @Nullable GuildSettings settings) {
         Set<Member> members = new LinkedHashSet<>();
-        members.addAll(getChannelContext(message, maxAge));
-        members.addAll(getVoiceContext(message, maxAge));
+        members.addAll(getChannelContext(message, settings));
+        members.addAll(getVoiceContext(message, settings));
         return members;
     }
 }
