+++ src/main/java/de/chojo/repbot/manager/ReputationService.java
@@ -6,7 +6,7 @@ import de.chojo.repbot.config.elements.MagicImage;
 import de.chojo.repbot.data.GuildData;
 import de.chojo.repbot.data.ReputationData;
 import de.chojo.repbot.data.wrapper.GuildSettings;
-import de.chojo.repbot.util.ContextResolver;
+import de.chojo.repbot.analyzer.ContextResolver;
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.entities.Guild;
 import net.dv8tion.jda.api.entities.Member;
